Fix not being able to inherit from DBusMockObject.
Don't use super(self.__class__, self).__init__() since this implies that
it will call the __init__() function of our class instead of the base
class. This fails if we have something like this:

class A(DBusMockObject):
    def __init__(self, a):
        DBusMockObject.__init__(self, ...)

In this example, DBusMockObject's __init__() would end up calling A's
__init__(). Not really what we want.

The better fix would be to use just super(), but that only works with Python 3.

Thanks Lucas De Marchi!