Add support for the D-Bus ObjectManager interface
An optional is_object_manager parameter to the DBusMockObject
constructor allows objects which implement the ObjectManager
interface to be exposed on the bus. This provides some level of
convenience by eliminating the need to describe this standard
interface.

Note that InterfacesAdded and InterfacesRemoved signals are *not*
emitted automatically when objects are added and removed, as a
single such signal could potentially correspond with a series of
python-dbusmock method calls. Support for them can be added in
future if python-dbusmock grows a method for atomically adding
multiple objects, each with several interfaces (and the same
for removals).