Add a Reset method to DBusMockObject
This method is accessible over D-Bus and will reset that object’s
state (including removing all its sub-objects) as if the object had
been destroyed and re-created. If the object was originally created
with a template, its methods and interfaces (etc.) will be re-loaded
from the template.

This is intended for use at the end of test suites to reset state
before the next unit test is run, as an alternative to killing
python-dbusmock and re-starting it (which is more expensive).