Allow Python functions to be passed to AddMethod from Python code
If running a test from within Python, it’s convenient to be able to
add methods by passing in a Python function handle, rather than
just a string.

This can also be used from within template files, which allows for
more complex behaviour than can be fitted in a string of inline code.

This is necessary to be able to add methods to objects which are
not the main object in a template file.