Rearchitect Transfer management in the bluez5-obex template
The pbap-client program expects the Transfer.Filename property to always be
available and always set to the transfer filename. This makes sense, so
rearchitect Transfer handling so that the filename is specified in the
TransferCreated mock signal, and the test driver code writes to this file
throughout the transfer, calling UpdateStatus on the transfer for each write
it wishes to make known to the software under test.