Add unit tests for the ‘bluez5-obex’ template
These cover all functionality of the test in a single test, since the
template is fairly limited at the moment.

The tests use the pbap-client test program which comes with BlueZ. This
program is not installed by BlueZ, so must be manually copied into the
user’s PATH before the bluez5-obex tests will run.

This seems to be unavoidable, due to the lack of any other available
programs which use Bluetooth PBAP.