Fix logging of unicode strings in Python 2
In Python2 one can't call `str(x)` if `x` is an unicode object, this
causes a `UnicodeEncodeError`.

Log unicode objects with their `repr` instead. This code path never gets
run for Python 3 as unicode is an alias for str there.

Change the log writing to be binary, so that UTF-8 strings properly work
regardless of the environment's locale.

Fixes #26