tests: Use nosetests directly in Fedora runner
This avoids having to copy the source tree, as that does not involve
creating the egg-info directory. `./setup.py test` is apparently also
deprecated:

    WARNING: Testing via this command is deprecated and will be removed in a future version.
    Users looking for a generic test entry point independent of test runner are encouraged to use tox.

Call nosetests directly. This is also a much more convenient way how to
iterate manually, as you can make changes on the host and directly test
them in the container.