templates: Add systemd template
The template works both when used as a system or a user service manager,
and shares a common implementation. It currently implements mocked
versions of StartUnit, StartTransientUnit and StopUnit. It imitates
systemd enough to make gnome-shell happy, thinking it managed to launch
the units it asked to have launched. To launch either the session or
system version of the tempalte, use system_bus named argument in
spawn_server_template().

The same added test is is run twice, one for the session bus and one for
the system bus. It uses the systemd D-Bus API directly instead of e.g.
systemctl, as systemctl uses things we can't easily mock, i.e. that
doesn't pass via the systemd D-Bus API.