template/accounts_service: Add accounts service mock implementation (#68)
It allows to create fake users to be added to the accountsservice,
implementing all the service methods.

All the user properties can be simply mocked calling the adduser mock
method or during initialization via parameters.

The tests may still fail as per some AccountsService changes that are
not yet released. So, the required dependencies do not get installed into
tests yet.