Ensure method calls from templates are logged
For interfaces loaded via a template, calls to static methods were never
logged and the MethodCalled signal was never emitted. Fix this by
wrapping the function from the module in what is a "standard" Python
function debug decorator, except that it adds the calls to the log
instead.

This is a bit awkward since the first argument is always self - we need
to pass that on to the function but we don't want to log it.