mockobject: allow sending signals with extra details
Add a new DBus method EmitSignalDetailed that also takes an a{sv} of
magic keywords for details to be passed down. The only supported detail
right now is "destination". This allows us to send signals to one
specific target even if they don't have a match rule setup (used by
libportal for example).