Bring back dbusmock.__version__
autotools's version checks or upower's test suite rely on this, so this
is API. Add a unit test to ensure it exists.

This should ideally be generated automatically [1], but that does not
work as long as packit, rpm builds etc. don't get along with pyproject.

So for the time being, move the static version back to
dbusmock/__init__.py, and read it from setup.py instead.

[1] https://github.com/martinpitt/python-dbusmock/pull/140