setup.py: Fix ModuleNotFoundError
Commit 842d2506fb71 broke `pip install` and also some variants of
PEP-517 build tools: `.` is not in their default `sys.path` any more.
Merely adding that to the search path is not sufficient: building an
sdist in a venv is tripping over the missing `import dbus`, as
dbus-python is just a runtime dependency, not a build-time one.

So instead of importing `dbusmock.__version__` in setup.py, ask
setuptools to dynamically generate the version in a declarative way, in
setup.cfg.

Many thanks to Michał Górny for analyzing this and proposing this
solution!

Fixes #142