Explicitly opt into static code tests
Running them merely when they are installed is a trap: We know that our
code does not satisfy the latest pylint release. The static code tests
are for upstream development, not for downstream or user-side testing.

So only run them when `$TEST_CODE` is set. Plumb that through the
container and the script, and set it in our GitHub workflow.

Fixes #155