mockobject: allow sending signals from a specific path
While we can easily mock.AddObject() an object with a different path,
sending a signal from that object is more complicated. And even there it
doesn't cover the case where we want to fake a signal from an object
created by someone else.

So, let's fix this by adding a path argument to EmitSignalDetailed.