Add initial pytest support
Add a new dbusmock.pytest_fixtures module which exports an initial set
of fixtures. These mostly just wrap DbusTestCase into a fixture, plus
some convenience functions for starting a system or session bus. In the
future we may have more fixtures to spawn templates, etc.

Ensure that all of our unit tests run with pytest. As that is expensive,
only run it for TEST_CODE=1 in Fedora stable. On the others, only run
the basic TestAPI checks and the fixtures checks.

This was co-developed with Peter Hutterer @whot, thank you!