testcase: reset the datadir on the DBusTestCaseClass
On the run of a derived test class, cls._DBusTestCase__datadir calls up
into the parent class for the datadir and removes it. The
subsequent statement cls._DBusTestCase__datadir = None sets it on the
derived class, not on DBusTestCase which remains at the original value.

Doesn't matter too much since the next derived test class will repeat
that process but we leave the parent datadir set despite having removed
it which seems like a bad idea.

See the comment in get_services_dir(), the datadir is always on
the DBusTestCase class, so let's mirror that behavior here.