testcase: start the bus on our class, not on the parent
A slightly weird side-effect here: start_dbus was called on the
DBusTestCase class - thus the session/system bus pid was set on that
class.

But tearDownClass() was called with the derived class from the test, so
while the getters for variables resolved correctly,
   setattr(cls, f"{bus_type}_bus_pid", None)
would set TestFoo.session_bus_pid = None and leave
DBusTestCase.session_bus_pid at whatever value it had before.

Fix this and add an assert to ensure we never start a session/system bus
twice in the same test class.