testcase: assert our parent class never gets bus PIDs
This is a potential break but it's unlikely to affect any old-style
unittest classes since they would be subclasses by default. It may
affect new-style pytest classes, if any exist, but OTOH it
can catch confusing errors like setting a variable on the DBusTestCase
class but then unsetting it on the derived class during a test.