testcase: replace the classmethods with a proper object
This patch introduces the BusType class which handles the system/session
buss differences including whatever is independent of who started the
actual daemon such as getting a connection.

It moves spawn_server and spawn_server_template to standalone functions
since those do not have any dependency on anything else in
DBusTestCase.

It adds the PrivateDBus class which wraps a single instance of a DBus
daemon, with a context manager so we can do things like:

    with PrivateDBus(BusType.SESSION) as bus:
        ....

and have it self clean up automatically. Unlike the previous approach,
we don't just fork in the background and kill via PID, we keep a
reference to the process and kill it via Python's API.

Finally, it changes the DBusTestCase class to use a single instance of
those objects as its DBus server when created with start_session_bus()
or start_system_bus(). Most of that class is now a wrapper around the
PrivateDBus object (or the BusType object where appropriate).

PrivateDBus also wraps some functions to make it look more like a
DBusTestCase object to keep compatibility with the current pytest
behavior.

With all that, we can now return a PrivateDBus object from the pytest
fixtures and have better insulation than before where the DBusTestCase
class was responsible for singleton handling.