power_profiles_daemon: Move to org.freedesktop.UPower.PowerProfiles
Version 0.20 moved to D-Bus name `org.freedesktop.UPower.PowerProfiles`,
and corresponding changes to the object path/interface:
https://gitlab.freedesktop.org/upower/power-profiles-daemon/-/releases/0.20

The real daemon also still listens to the old name, but we can't do that
in dbusmock: BUS_NAME and friends cannot be parameterized, as their
loading and instantiation happens before calling the `load()` function.

`powerprofilesctl` now only looks at the new name, so we need to adjust
the template. The new version is in Fedora 39 to rawhide and also in
Debian testing, so for ongoing development it's more useful to present
the current API.