Add initial ModemManager mock
We mock the daemon and allow to add a modem (which also adds a SIM)
and a minimal Voice interface.

This is enough for `mmcli -m any` and `mmcli -i any` to work, for shells
like phosh to display mobile broadband information and for calls thinking
it's seeing a voice capable modem (thus enabling call functionality).