bluez5: Implement LE Advertising Manager API
Expose the `org.bluez.LEAdvertisingManager1` interface on adapters created
via the `AddAdapter` method. The included stubs of `RegisterAdvertisement`
and `UnregisterAdvertisement` implement very basic behaviour only and do not
fully emulate the behaviour of BlueZ.

The following cases are handled:

- Tracking registered paths to prevent duplicate registrations and to
  prevent unregistration of unknown advertisements.
- Tracking number of active advertisements to prevent more than the
  number of supported advertisements being registered, and to maintain the
  `ActiveInstances` and `SupportedInstances` properties.

The implementation does not attempt to access, or interact with the DBus object
at the given path.

This is enough to get `bluetoothctl` to detect that the Bluetooth adapter
is capable of BLE advertising, list the supported advertising
features, and register/unregister advertisements.

This interface can be toggled with the parameter `enable_advertise_api`
when spawning the bluez5 template.