bluez5: Implement Advertisement Monitor Manager API
Expose the `org.bluez.AdvertisementMonitorManager1` interface on adapters
created via the `AddAdapter` method. The included stubs of `RegisterMonitor`
and `UnregisterMonitor` implement very basic behaviour only and do not
fully emulate the behaviour of BlueZ.

The following cases are handled:

- Tracking registered paths to prevent duplicate registrations and to prevent
  unregistration of unknown monitors.

The implementation does not attempt to access, or interact with the DBus object
at the given path.

This is enough to get `bluetoothctl` to detect that the Bluetooth adapter
is capable of passive scanning, and list the supported advertising
monitor features.

This interface can be toggled with the parameter `enable_monitor_api`
when spawning the bluez5 template.