Fix dbus-daemon startup race condition
With the previous polling loop there was a short window where the socket
existed, but dbus-daemon didn't listen to it yet. Fix this by using
`--print-pid` and waiting for the result. dbus-daemon guarantees that
this will only be printed after it is ready to accept connections.

Thanks to Simon McVittie for finding this and proposing the solution!
(Ironically, the dead code removed in the previous commit did it right.
But avoid forking here.)

https://bugs.debian.org/1109272